### 2023-02-03 cc7fdd1
finish norm metrics testing
