### 2023-02-03 b8bcb4e
start norm plot testing
