+++ tests/testthat/test-normalization_plotting.R
@@ -1,0 +1,8 @@
+test_that("eval_pn_metric_for_plot output has proper structure", {
+
+})
+
+
+# At the moment, not going to do testing of the
+# various plotting functions, which should all output ggplot objects
+# and testing all of them for correctness would be tricky.
