### 2023-02-03 92c4d60
normalization_plotting testing
