### 2023-02-03 243f0b0
speed up normalization_plotting tests
