+++ tests/testthat/fixtures/create_norm-report_input.R
@@ -10,14 +10,14 @@ test_metadata <- data.frame(sample_ID = paste0("sample", 1:6),
                                           rep("treatment", 3)))
 rownames(test_metadata) <- test_metadata$sample_ID
 set.seed(42)
-test_data <- data.frame(a = as.integer(runif(n = 2000, min = 10000, max = 20000)),
-                        b = as.integer(runif(n = 2000, min = 10000, max = 20000)),
-                        c = as.integer(runif(n = 2000, min = 10000, max = 20000)),
-                        d = as.integer(runif(n = 2000, min = 10000, max = 20000)),
-                        e = as.integer(runif(n = 2000, min = 10000, max = 20000)),
-                        f = as.integer(runif(n = 2000, min = 10000, max = 20000)))
+test_data <- data.frame(a = as.integer(runif(n = 100, min = 10000, max = 20000)),
+                        b = as.integer(runif(n = 100, min = 10000, max = 20000)),
+                        c = as.integer(runif(n = 100, min = 10000, max = 20000)),
+                        d = as.integer(runif(n = 100, min = 10000, max = 20000)),
+                        e = as.integer(runif(n = 100, min = 10000, max = 20000)),
+                        f = as.integer(runif(n = 100, min = 10000, max = 20000)))
 colnames(test_data) <- test_metadata$sample_ID
-test_annotation <- data.frame(uniprot_id = paste0("protein", 1:2000))
+test_annotation <- data.frame(uniprot_id = paste0("protein", 1:100))
 input <- DAList(data = test_data,
                 annotation = test_annotation,
                 metadata = test_metadata)
