### 2023-02-03 ca206f8
tests for write_nrom_report
