### 2023-02-07 04994fd
all tests for write_qc_report
