### 2023-02-10 e8f1ef8
tests for write_limma_tables
