+++ tests/testthat/test-limma_tables.R
@@ -1,0 +1,233 @@
+# A lot of these functions are tricky to test,
+# the base code could probably use a re-factor.
+# But, tried to test as much as possible.
+
+test_that("make_excel_hyperlinks adds hyperlinks to col", {
+  df <- data.frame(x = c("A",
+                         "B",
+                         "C",
+                         NA))
+  expect <- data.frame(x = c("HYPERLINK(\"test.com/A\", \"A\")",
+                             "HYPERLINK(\"test.com/B\", \"B\")",
+                             "HYPERLINK(\"test.com/C\", \"C\")",
+                             NA))
+  class(expect[,"x"])  <- "formula"
+
+  expect_equal(make_excel_hyperlinks(df, url.col = 1, url = "test.com/"), expect)
+
+})
+
+
+test_that("write_limma_excel returns consistent output", {
+
+  input <- readRDS(test_path("fixtures", "final_output_input.rds"))
+
+  on.exit(unlink("temp.xlsx"), add = T)
+
+  suppressMessages(
+      output <- write_limma_excel(filename = "temp.xlsx",
+                        statlist = input$results,
+                        annotation = input$annotation,
+                        data = input$data,
+                        norm.method = input$tags$norm_method,
+                        pval_thresh = input$tags$DA_criteria$pval_thresh,
+                        lfc_thresh = input$tags$DA_criteria$lfc_thresh,
+                        add_filter = T)
+  )
+
+  # Track the the worksheet output, should be consistent
+  expect_snapshot(output$wb$worksheets)
+
+})
+
+test_that("create_combined_results returns expected output", {
+
+  input <- readRDS(test_path("fixtures", "final_output_input.rds"))
+
+  a <- input$results$control
+  colnames(a) <- paste(colnames(a), "control", sep = "_")
+  b <- input$results$treatment
+  colnames(b) <- paste(colnames(b), "treatment", sep = "_")
+
+  expected_output <- cbind(input$annotation,
+                           input$data,
+                           a,
+                           b)
+
+  expect_equal(create_combined_results(annotation = input$annotation,
+                                       data = input$data,
+                                       statlist = input$results),
+              expected_output)
+})
+
+
+test_that("summarize_contrast_DA returns expected output", {
+
+  input <- readRDS(test_path("fixtures", "final_output_input.rds"))
+  # Add in a mix of all significance types in the fake data
+  input$results$control$sig.PVal <- c(rep(-1, 4),
+                                      rep(0, 3),
+                                      rep(1, 3))
+  input$results$control$sig.FDR <- c(rep(-1, 3),
+                                      rep(0, 4),
+                                      rep(1, 3))
+  input$results$treatment$sig.PVal <- c(rep(-1, 4),
+                                        rep(0, 3),
+                                        rep(1, 3))
+  input$results$treatment$sig.FDR <- c(rep(-1, 3),
+                                       rep(0, 4),
+                                       rep(1, 2),
+                                       NA)
+
+  control_expect <- data.frame(contrast = "control",
+                               type = c("down", "nonsig", "up"),
+                               sig.PVal = as.character(c(4,3,3)),
+                               sig.FDR = as.character(c(3,4,3)))
+  treatment_expect <- data.frame(contrast = "treatment",
+                               type = c("down", "nonsig", "up"),
+                               sig.PVal = as.character(c(4,3,3)),
+                               sig.FDR = as.character(c(3,4,2)))
+
+  expect_equal(summarize_contrast_DA("control", input$results), control_expect)
[…]
