+++ tests/testthat/test-limma_report.R
@@ -1,0 +1,273 @@
+test_that("prep_plot_model_data outputs the proper data and format", {
+  input <- readRDS(test_path("fixtures", "final_output_input.rds"))
+
+  # data frame with expected colnames,
+  # same number of rows as data,
+  # and no missing data in the sig cols
+
+  out_control <- prep_plot_model_data(model_results = input$results,
+                                      contrast = "control")
+
+  out_control <- prep_plot_model_data(model_results = input$results,
+                                      contrast = "treatment")
+
+  expect_cols <- c("logFC", "CI.L", "CI.R", "AveExpr", "t", "B", "P.Value",
+                   "adj.P.Val", "sig.PVal", "sig.FDR", "P value",
+                   "Adjusted P value", "negLog10rawP", "negLog10adjP",
+                   "sig.pval.fct", "sig.FDR.fct")
+
+
+
+  for (contrast in c("control", "treatment")) {
+    output <- prep_plot_model_data(model_results = input$results,
+                                   contrast = contrast)
+
+    expect_equal(colnames(output), expect_cols)
+    expect_equal(nrow(output), nrow(input$data))
+    expect_equal(sum(is.na(output$sig.PVal)), 0)
+    expect_equal(sum(is.na(output$FDR)), 0)
+  }
+})
+
+# don't test ggplot functions
+
+
+
+# tests for write_limma_plots should be similar to
+# limma tables tests
+
+test_that("write_limma_plots gives error when there are no results", {
+
+  input <- readRDS(test_path("fixtures", "norm_report_input.rds"))
+
+
+  expect_error(write_limma_plots(input, overwrite = T), "does not have results")
+})
+
+test_that("write_limma_plots checks grouping column", {
+
+  input <- readRDS(test_path("fixtures", "final_output_input.rds"))
+
+  # must specify grouping col
+  expect_error(write_limma_plots(input), "cannot be empty")
+
+  # can't specify more than one
+  expect_error(write_limma_plots(input, grouping_column = c("group", "batch")),
+               "Only specify one")
+
+  # must be present in the data
+  expect_error(write_limma_plots(input, grouping_column = "xxx"),
+               "not found")
+
+})
+
+test_that("write_limma_plots checks table columns", {
+
+  input <- readRDS(test_path("fixtures", "final_output_input.rds"))
+
+  expect_error(
+    write_limma_plots(input,
+                      grouping_column = "treatment",
+                      table_columns = c("xxx", "yyy")),
+    "not found in annotation"
+  )
+})
+
+test_that("write_limma_plots checks title column", {
+
+  input <- readRDS(test_path("fixtures", "final_output_input.rds"))
+
+  # Must be of length 1
+  expect_error(
+    write_limma_plots(input,
+                      grouping_column = "treatment",
+                      title_column = c("xxx", "yyy")),
+    "does not equal 1"
+  )
+
+
+  # must be present
+  expect_error(
+    write_limma_plots(input,
+                      grouping_column = "treatment",
+                      title_column = c("xxx")),
+    "not found in annotation"
+  )
+
+  # values must be unique
+  input$annotation$uniprot_id[2] <- input$annotation$uniprot_id[1]
+  expect_error(
+    write_limma_plots(input,
+                      grouping_column = "treatment"),
+    "not unique"
+  )
+})
+
+test_that("write_limma_plots check height and width args", {
+  input <- readRDS(test_path("fixtures", "final_output_input.rds"))
+
+  expect_error(
+    write_limma_plots(input,
+                      grouping_column = "treatment",
+                      height = "xxx"),
+    "numeric value greater"
+  )
+
+  expect_error(
+    write_limma_plots(input,
+                      grouping_column = "treatment",
+                      height = -1),
+    "numeric value greater"
+  )
+
+  expect_error(
[…]
