### 2023-02-21 #5184 @editorialbot
[REVIEW]: proteoDA: a package for quantitative proteomics
**Submitting author:** @ByrumLab (<a href="http://orcid.org/0000-0002-1783-3610">Stephanie D. Byrum</a>)
**Repository:** https://github.com/ByrumLab/proteoDA
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.0
