### 2023-02-26 #5184 @MohmedSoudy
Hi @ByrumLab, Thank you for that great software package. The paper is well-written and the package is well-documented, however, I have some comments that may increase the package's feasibility and usability
-  State of the field: I believe that this point should be highlighted in the paper to show the significance of the package and the contribution performed by the authors.
- write_qc_report function: The names of the samples in the heat map should fit in the graph, and the legend position may be at the right and should be bigger to be more readable
- write_limma_plots function: The user should select the p-value adjustment method on the plot and re-plot, and you can plot the box plot for each protein with the point plot for better visualization of the group mean/median and the outliers
