### 2023-02-28 cf8d930
update QC report to address comments
truncate sample IDs to 20 char so they don't overlap legends,. Increase legend size of all plots. For >50 samples, dynamically size (with a minimum) QC plots.  Add some handling of long sample ids: truncates, with warning, IDs to 20 characters (errors if no longer unique) and, where needed, modifies margins to accommodate 20 character labels.
