### 2023-03-02 636ed27
passing full results to javascript, force MD plot for now
