+++ R/glimma_fxns.R
@@ -1,33 +1,30 @@
-uams_glimmaXY <- function (x, y, xlab = "x", ylab = "y", dge = NULL, counts = dge$counts,
-                           groups = dge$samples$group, status = rep(0, length(x)), anno = NULL,
-                           display.columns = NULL, status.cols = c("#1052bd", "silver",
-                                                                   "#cc212f"), sample.cols = NULL, main = "XY Plot", html = NULL,
-                           width = 920, height = 920)
-{
-  if (length(x) != length(y))
-    stop("Error: x and y args must have the same length.")
-  table <- data.frame(signif(x, digits = 4), signif(y, digits = 4))
-  colnames(table) <- c(xlab, ylab)
-  if (!is.null(counts)) {
-    table <- cbind(gene = rownames(counts), table)
-  }
-  else if (!is.null(rownames(x))) {
-    table <- cbind(gene = rownames(x), table)
-  }
-  else if (!is.null(rownames(y))) {
-    table <- cbind(gene = rownames(y), table)
-  }
-  else {
-    table <- cbind(gene = seq_along(x), table)
-  }
+uams_glimmaXY <- function(model_data,
+                          counts,
+                          groups,
+                          status,
+                          anno,
+                          display.columns,
+                          status.cols,
+                          sample.cols,
+                          main = "",
+                          html = NULL,
+                          width = 920,
+                          height = 920) {
+
+  # The model table should be the output from
+  # prep_plot_model_data.
+  # Originally these were rounded to 4 digits, but maybe don't need to do that?
+  # could if we really need to save space
+  table <- model_data
+  table$gene <- rownames(model_data)
+
   xData <- UAMS_buildXYData(table, status, main, display.columns,
-                            anno, counts, xlab, ylab, status.cols, sample.cols, groups)
+                            anno, counts, status.cols, sample.cols, groups)
   return(UAMS_glimmaXYWidget(xData, width, height, html))
 }
 
 
-UAMS_glimmaXYWidget <- function (xData, width, height, html)
-{
+UAMS_glimmaXYWidget <- function (xData, width, height, html) {
   widget <- htmlwidgets::createWidget(name = "glimmaXY", xData, package = "proteoDA",
                                       width = width, height = height, elementId = NULL,
                                       sizingPolicy = htmlwidgets::sizingPolicy(defaultWidth = width,
@@ -43,51 +40,43 @@ UAMS_glimmaXYWidget <- function (xData, width, height, html)
 }
 
 
-UAMS_buildXYData <- function (table, status, main, display.columns, anno, counts,
-                              xlab, ylab, status.cols, sample.cols, groups)
-{
-  if (is.null(counts)) {
-    counts <- -1
-    level <- NULL
+UAMS_buildXYData <- function(table, status, main, display.columns, anno, counts,
+                             status.cols, sample.cols, groups) {
+
+  counts <- data.frame(counts)
+  if (is.null(groups)) {
+    groups <- factor("group")
   }
   else {
-    counts <- data.frame(counts)
-    if (is.null(groups)) {
-      groups <- factor("group")
-    }
-    else {
-      if (ncol(counts) != length(groups))
-        stop("Length of groups must be equal to the number of columns in counts.\n")
-    }
-    level <- levels(groups)
-    groups <- data.frame(group = groups)
-    groups <- cbind(groups, sample = colnames(counts))
+    if (ncol(counts) != length(groups))
+      stop("Length of groups must be equal to the number of columns in counts.\n")
   }
+
+  level <- levels(groups)
+  groups <- data.frame(group = groups)
+  groups <- cbind(groups, sample = colnames(counts))
+
[…]
