### 2023-03-03 dadb564
select axes with dropdown menus
