+++ inst/htmlwidgets/lib/vega-tooltip/vega-tooltip.js
@@ -523,4 +523,3 @@
     Object.defineProperty(exports, '__esModule', { value: true });
 
 })));
-//# sourceMappingURL=vega-tooltip.js.map
