### 2023-03-06 6a2704a
remove unneeded .Rmd template, minor code comments
