+++ inst/report_templates/glimma_xy_plot.Rmd
@@ -1,49 +1,0 @@
-## `r noquote({{plot}})`
-
-```{r, results = "asis"}
-if ({{plot}} == "Volcano raw P") {
-  x <- data$logFC
-  y <- data$negLog10rawP
-  xlab <- "logFC"
-  ylab <- "negLog10rawP"
-  status <- data$sig.PVal
-  cols_to_display <- c(internal_table_columns, "logFC", "p", "adjusted_p")
-} else if ({{plot}} == "Volcano adjusted P") {
-  x <- data$logFC
-  y <- data$negLog10adjP
-  xlab <- "logFC"
-  ylab <- "negLog10adjP"
-  status <- data$sig.FDR
-  cols_to_display <- c(internal_table_columns, "logFC", "p", "adjusted_p")
-} else if ({{plot}} == "MD raw P") {
-  x <- data$AveExpr
-  y <- data$logFC
-  xlab <- "AveExpr"
-  ylab <- "logFC"
-  status <- data$sig.PVal
-  cols_to_display <- c(internal_table_columns, "AveExpr", "logFC", "p", "adjusted_p")
-} else if ({{plot}} == "MD adjusted P") {
-  x <- data$AveExpr
-  y <- data$logFC
-  xlab <- "AveExpr"
-  ylab <- "logFC"
-  status <- data$sig.FDR
-  cols_to_display <- c(internal_table_columns, "AveExpr", "logFC", "p", "adjusted_p")
-} else {
-  stop("Invalid plot option")
-}
-
-uams_glimmaXY(
-        x = x,
-        y = y,
-        counts = counts,
-        groups = groups,
-        sample.cols = unname(colorGroup(groups)[groups]),
-        status.cols = c("#00bfff", "#858585", "#ff3030"),
-        status = status,
-        anno = anno,
-        display.columns = cols_to_display,
-        main = " ",
-        width = width,
-        height = height)
-```
