### 2023-03-06 26e339a
sync XY plot color with pval selector, improve axis titles
