### 2023-03-06 fd92ce7
add median lines to intensity plot
