+++ tests/testthat/test-limma_report.R
@@ -11,7 +11,7 @@ test_that("prep_plot_model_data outputs the proper data and format", {
   out_control <- prep_plot_model_data(model_results = input$results,
                                       contrast = "treatment")
 
-  expect_cols <- c("logFC", "CI.L", "CI.R", "AveExpr", "t", "B", "P.Value",
+  expect_cols <- c("logFC", "CI.L", "CI.R", "average_intensity", "t", "B", "P.Value",
                    "adj.P.Val", "sig.PVal", "sig.FDR", "P value",
                    "Adjusted P value", "negLog10rawP", "negLog10adjP",
                    "sig.pval.fct", "sig.FDR.fct")
