### 2023-03-07 d696112
improve error message when sample IDs are not unique after truncation
