### 2023-03-07 9f8446e
dynamically sizes width of median line in abundance plot
