+++ tests/test_workflow.R
@@ -314,10 +314,6 @@ write_limma_tables(results_higgs,
 
 # testing report making ---------------------------------------------------
 
-write_limma_plots(results_reb,
-                  grouping_column = "group",
-                  output_dir = "reb_s3obj_pre_edits",
-                  title_column = "Accession.Number", overwrite = T)
 
 write_limma_plots(results_reb,
                   grouping_column = "group",
