+++ R/glimma_fxns.R
@@ -84,7 +84,7 @@ UAMS_buildXYData <- function(table, status, main, display.columns, anno, counts,
       groups = groups,
       levels = level,
       expCols = colnames(groups),
-      numUniqueGroups = length(unique(groups)),
+      numUniqueGroups = length(unique(groups$group)),
       annoCols = if (is.null(anno)) {-1} else {colnames(anno)},
       statusColours = status.cols,
       sampleColours = if (is.null(sample.cols)) {-1} else {sample.cols},
