### 2023-03-08 1d3a820
switch from medians to means
