### 2023-03-08 bfe8dfd
bugfix for selection with means
the console logging was throwing an error that broke selection
