+++ inst/htmlwidgets/glimmaXY.js
@@ -375,10 +375,10 @@ function updateExpressionPlot(countsRow, data, geneName) {
   data.expressionView.runAsync();
   updateAxisMessage(data);
   // log the median data set
-  console.log("Table data");
-  console.log(JSON.parse(JSON.stringify(data.expressionView.data("table"))));
-  console.log("Median data");
-  console.log(data.expressionView.data("medians"));
+  //console.log("Table data");
+  //console.log(JSON.parse(JSON.stringify(data.expressionView.data("table"))));
+  //console.log("Median data");
+  //console.log(data.expressionView.data("medians"));
 
 }
 
