### 2023-03-08 c396a2b
fix an issue where title columns with special characters failed to match
Bypass the old way of setting new unique keys/rownames. Just make a new internal title column and pass that value as the title signal.
