+++ paper/references.bib
@@ -156,3 +156,49 @@ year = {2022}
     eprint = {https://academic.oup.com/nar/article-pdf/50/D1/D543/42057564/gkab1038.pdf},
 }
 
+@article {Ranathunge2022,
+	author = {Ranathunge, Chathurani and Patel, Sagar S. and Pinky, Lubna and Correll, Vanessa L. and Chen, Shimin and Semmes, O. John and 
+Armstrong, Robert K. and Combs, C. Donald and Nyalwidhe, Julius O.},
+	title = {promor: a comprehensive R package for label-free proteomics data analysis and predictive modeling},
+	elocation-id = {2022.08.17.503867},
+	year = {2023},
+	doi = {10.1101/2022.08.17.503867},
+	publisher = {Cold Spring Harbor Laboratory},
+	abstract = {Summary We present promor, a comprehensive, user-friendly R package that streamlines label-free (LFQ) proteomics data 
+analysis and building machine learning-based predictive models with top protein candidates.Availability and implementation promor is freely 
+available as an open source R package on the Comprehensive R Archive Network (CRAN)(https://CRAN.R-project.org/package=promor) and distributed 
+under the Lesser General Public License (version 2.1 or later). Development version of promor is maintained on GitHub 
+(https://github.com/caranathunge/promor) and additional documentation and tutorials are provided on the package website 
+(https://caranathunge.github.io/promor/).Contact caranathunge86{at}gmail.comCompeting Interest StatementThe authors have declared no competing 
+interest.},
+	URL = {https://www.biorxiv.org/content/early/2023/01/16/2022.08.17.503867},
+	eprint = {https://www.biorxiv.org/content/early/2023/01/16/2022.08.17.503867.full.pdf},
+	journal = {bioRxiv}
+}
+
+ @article {Wolski2022,
+	author = {Wolski, Witold E. and Nanni, Paolo and Grossmann, Jonas and d{\textquoteright}Errico, Maria and Schlapbach, Ralph and Panse, 
+Christian},
+	title = {prolfqua: A Comprehensive R-package for Proteomics Differential Expression Analysis},
+	elocation-id = {2022.06.07.494524},
+	year = {2022},
+	doi = {10.1101/2022.06.07.494524},
+	publisher = {Cold Spring Harbor Laboratory},
+	abstract = {Mass spectrometry is widely used for quantitative proteomics studies, relative protein quantification, and differential 
+expression analysis of proteins. Nevertheless, there is a need for a flexible and easy-to-use application programming interface in R that 
+transparently supports a variety of well principled statistical procedures. The prolfqua package can model simple experimental designs with a 
+single explanatory variable and complex experiments with multiple factors and hypothesis testing. It integrates essential steps of the mass 
+spectrometry-based differential expression analysis workflow: quality control, data normalization, protein aggregation, statistical modeling, 
+hypothesis testing, and sample size estimation. The application programmer interface strives to be clear, predictable, discoverable, and 
+consistent to make proteomics data analysis easy and exciting. Furthermore, the package implements benchmark functionality that can help to 
+compare data acquisition, data preprocessing, or data modeling methods using a gold standard dataset. Finally, we show that the implemented 
+methods allow sensitive and specific differential expression analysis. The prolfqua R package is available on GitHub 
+https://github.com/fgcz/prolfqua, distributed under the MIT licence, and runs on all platforms supported by the R free software environment for 
+statistical computing and graphics.Competing Interest StatementThe authors have declared no competing interest.AbbreviationsExplainationAUCArea 
[…]
