+++ paper/paper.md
@@ -57,7 +57,8 @@ bibliography: references.bib
 
 # State of the field
 
-Quantitative proteomics analysis is growing in popularity with the advancement of mass spectrometer instrumentation and software that allows for the identification of more and more proteins from complex sample mixtures [@Arora2019; @Deutsch2023; @Perez-Riverol2022]. Sequencing depths have increased from a few thousand to 10,000 proteins commonly identified from discovery experiments. Most proteomics mass spectrometry workflows follow the shotgun or bottom-up approach [@Tariq2021] in which a mixture of proteins is digested by an enzyme into peptides and then analyzed by LC MS/MS. Expertise in sample preparation, mass spectrometer parameters, experimental study designs, biostatistics, and programming languages are all needed in order to properly analyze and interpret proteomics data. ProteoDA was designed to make the process of analyzing bottom-up MS intensity-based data streamlined with flexible model designs. The functions in proteoDA allow a practical approach [@Schessner2022] for the evaluation of normalization methods, easy selection of limma model designs to account for batch effects or other factors, and interactive visualization of the results. In contrast to other comprehensive proteomics analysis packages, `proteoDA` delivers an interactive, sharable data product that allows domain researchers to visualize their results in a web browser with the need for dedicated server-side infrastructure [@Ranathunge2022; @Wolski2022].
+Quantitative proteomics analysis is growing in popularity with the advancement of mass spectrometer instrumentation and software that allows for the identification of more and more proteins from complex sample mixtures [@Arora2019; @Deutsch2023; @Perez-Riverol2022]. Sequencing depths have increased from a few thousand to 10,000 proteins commonly identified from discovery experiments. Most proteomics mass spectrometry workflows follow the shotgun or bottom-up approach [@Tariq2021] in which a mixture of proteins is digested by an enzyme into peptides and then analyzed by LC MS/MS. Expertise in sample preparation, mass spectrometer parameters, experimental study designs, biostatistics, and programming languages are all needed in order to properly analyze and interpret proteomics data. ProteoDA was designed to make the process of analyzing bottom-up MS intensity-based data streamlined with flexible model designs. The functions in proteoDA allow a practical approach [@Schessner2022] for the evaluation of normalization methods, easy selection of limma model designs to account for batch effects or other factors, and interactive visualization of the results. In contrast to other comprehensive proteomics analysis packages, `proteoDA` delivers an interactive, sharable data product that allows domain researchers to visualize their results in a web browser without the need for dedicated server-side infrastructure [@Ranathunge2022; @Wolski2022].
+
 
 # Statement of need
 
