### 2023-03-09 a5c19c1
round numeric data columns in interactive report to 4 digits
makes interactive reports a little nicer to look at and slightly smaller
