### 2023-03-09 5449573
minor streamlining of limma_report code
