+++ R/limma_report.R
@@ -109,19 +109,18 @@ write_limma_plots <- function(DAList = NULL,
   # And set it
   if (is.null(grouping_column)) { # If no groups provided, abort
     cli::cli_abort("{.arg grouping_column} cannot be empty")
-  } else { # check that grouping column exists in the target data frame
-    if (length(grouping_column) != 1) {
-      cli::cli_abort(c("Length of {.arg grouping_column} does not equal 1",
-                       "i" = "Only specify one column name for {.arg grouping_column}"))
-
-    }
-    if (grouping_column %notin% colnames(DAList$metadata)) {
-      cli::cli_abort(c("Column {.arg {grouping_column}} not found in metadata of {.arg DAList}",
-                       "i" = "Check the column names with {.code colnames(DAList$metadata)}."))
-    }
-    # And set it
-    groups <- as.character(DAList$metadata[,grouping_column])
   }
+  if (length(grouping_column) != 1) {
+    cli::cli_abort(c("Length of {.arg grouping_column} does not equal 1",
+                     "i" = "Only specify one column name for {.arg grouping_column}"))
+  }
+  if (grouping_column %notin% colnames(DAList$metadata)) {
+    cli::cli_abort(c("Column {.arg {grouping_column}} not found in metadata of {.arg DAList}",
+                     "i" = "Check the column names with {.code colnames(DAList$metadata)}."))
+  }
+  # And set it
+  groups <- as.character(DAList$metadata[,grouping_column])
+
 
   # check that the table columns exist in the annotation
   if (!all(table_columns %in% colnames(DAList$annotation))) {
@@ -169,7 +168,6 @@ write_limma_plots <- function(DAList = NULL,
     cli::cli_inform("Setting output directory to current working directory:")
     cli::cli_inform("{.path {output_dir}}")
   }
-
 
 
   # Set up list of expected files, warn about overwriting if they exist
@@ -265,13 +263,16 @@ write_limma_plots <- function(DAList = NULL,
             to = "report_template.Rmd", overwrite = T)
 
   # once we create the files, ensure they're deleted if there's an error below
-  on.exit(expr = {
-    cli::cli_inform("Removing temporary files from {.path {output_dir}}")
-    unlink(c("logo_higherres.png", "report_template.Rmd", tmp_subdir), recursive = T, expand = F)
-  }, add = T, after = F)
+  on.exit(
+    expr = {
+      cli::cli_inform("Removing temporary files from {.path {output_dir}}")
+      unlink(c("logo_higherres.png", "report_template.Rmd", tmp_subdir), recursive = T, expand = F)
+    },
+    add = T,
+    after = F
+  )
 
   # Prep to loop over contrasts
-
   contrast_count <- 1
   num_contrasts <- length(names(DAList$results))
 
@@ -284,11 +285,11 @@ write_limma_plots <- function(DAList = NULL,
     # the javascript will be mad. Replace with underscores.
     tbl_cols <- which(colnames(shared_anno) %in% table_columns)
     colnames(shared_anno)[tbl_cols] <- stringr::str_replace_all(colnames(shared_anno)[tbl_cols],
-                                                         "\\.",
-                                                         "_")
+                                                                "\\.",
+                                                                "_")
     internal_table_columns <- stringr::str_replace_all(table_columns,
-                                              "\\.",
-                                              "_")
+                                                       "\\.",
+                                                       "_")
   } else { # If no periods, can just use the provided table columns internally
     internal_table_columns <- table_columns
   }
[…]
