### 2023-03-09 262dacb
remove no-longer-needed JS from report template
