### 2023-03-09 331a2db
new data column to use for internal brushing, instead of "gene"
avoids a likely possible issue where users already have a "gene" column in their annotation. 

also does some streamlining of internal glimma fxns
