+++ R/glimma_fxns.R
@@ -7,19 +7,28 @@ uams_glimmaXY <- function(model_data,
                           status.cols,
                           sample.cols,
                           main = "",
-                          html = NULL,
                           width = 920,
                           height = 920) {
 
   # The model table should be the output from
   # prep_plot_model_data.
-  # Originally these were rounded to 4 digits, but maybe don't need to do that?
-  # could if we really need to save space
   table <- model_data
-  table$gene <- rownames(model_data)
+  table$internal_id_for_brushing <- rownames(model_data)
+  display.columns <- c("internal_id_for_brushing", display.columns)
 
-  # Need to change this to a for loop, I guess?
-  # or maybe lapply? regular apply is corercing to matrix and messing up the types
+  # Some possible mild argument checking?
+  # Though user will never really call this
+  if (length(status.cols) != 3) {
+    stop("status.cols\n arg must have exactly 3 elements for [downreg, notDE, upreg]")
+  }
+  if (is.null(groups)) {
+    groups <- factor("group")
+  }
+  else {
+    if (ncol(counts) != length(groups))
+      stop("Length of groups must be equal to the number of columns in counts.\n")
+  }
+
 
   # Round all numeric values to 4 digits, to make everything look nicer in tables
   table <- data.frame(lapply(table, FUN = function(col) {
@@ -39,31 +48,31 @@ uams_glimmaXY <- function(model_data,
     }
     output
   }))
-
+  # Counts is a matrix, can just round it
   counts <- round(counts, digits = 4)
 
 
-
-
+  # Build the data to pass to the htmlwidget
   xData <- UAMS_buildXYData(table, status, main, display.columns,
                             anno, counts, status.cols, sample.cols, groups)
-  return(UAMS_glimmaXYWidget(xData, width, height, html))
-}
 
+  # make the widget
+  widget <- htmlwidgets::createWidget(
+    name = "glimmaXY",
+    xData,
+    package = "proteoDA",
+    width = width,
+    height = height,
+    elementId = NULL,
+    sizingPolicy = htmlwidgets::sizingPolicy(
+      defaultWidth = width,
+      defaultHeight = height,
+      browser.fill = TRUE,
+      viewer.suppress = TRUE
+    )
+  )
 
-UAMS_glimmaXYWidget <- function (xData, width, height, html) {
-  widget <- htmlwidgets::createWidget(name = "glimmaXY", xData, package = "proteoDA",
-                                      width = width, height = height, elementId = NULL,
-                                      sizingPolicy = htmlwidgets::sizingPolicy(defaultWidth = width,
-                                                                               defaultHeight = height, browser.fill = TRUE, viewer.suppress = TRUE))
-  if (is.null(html)) {
-    return(widget)
-  }
-  else {
-    message("Saving widget...")
-    htmlwidgets::saveWidget(widget, file = html)
-    message(html, " generated.")
-  }
+  widget
 }
 
 
@@ -71,38 +80,16 @@ UAMS_buildXYData <- function(table, status, main, display.columns, anno, counts,
                              status.cols, sample.cols, groups) {
 
   counts <- data.frame(counts)
-  if (is.null(groups)) {
-    groups <- factor("group")
-  }
-  else {
-    if (ncol(counts) != length(groups))
-      stop("Length of groups must be equal to the number of columns in counts.\n")
-  }
-
   level <- levels(groups)
   groups <- data.frame(group = groups)
   groups <- cbind(groups, sample = colnames(counts))
 
-  status <- sapply(status, function(x) {
-    switch(as.character(x), `-1` = "downReg", `0` = "nonDE",
-           `1` = "upReg")
-  })
[…]
