### 2023-03-09 155af2c
major streamlining of internal R glimma fxns
consolidate into 1 function, avoid passing unnecessary info to the html widget
