+++ R/glimma_fxns.R
@@ -6,15 +6,8 @@ uams_glimmaXY <- function(model_data,
                           display.columns,
                           status.cols,
                           sample.cols,
-                          main = "",
-                          width = 920,
-                          height = 920) {
-
-  # The model table should be the output from
-  # prep_plot_model_data.
-  table <- model_data
-  table$internal_id_for_brushing <- rownames(model_data)
-  display.columns <- c("internal_id_for_brushing", display.columns)
+                          width,
+                          height) {
 
   # Some possible mild argument checking?
   # Though user will never really call this
@@ -29,9 +22,15 @@ uams_glimmaXY <- function(model_data,
       stop("Length of groups must be equal to the number of columns in counts.\n")
   }
 
+  # The model table should be the output from
+  # prep_plot_model_data.
+  table_for_widget <- cbind(model_data, anno)
+  table_for_widget$internal_id_for_brushing <- rownames(model_data)
+  display.columns <- c("internal_id_for_brushing", display.columns)
+
 
   # Round all numeric values to 4 digits, to make everything look nicer in tables
-  table <- data.frame(lapply(table, FUN = function(col) {
+  table_for_widget <- data.frame(lapply(table_for_widget, FUN = function(col) {
     if (is.numeric(col)) {
       output <- round(col, digits = 4)
     } else {
@@ -40,21 +39,40 @@ uams_glimmaXY <- function(model_data,
     output
   }))
 
-  anno <- data.frame(lapply(anno, FUN = function(col) {
-    if (is.numeric(col)) {
-      output <- round(col, digits = 4)
-    } else {
-      output <- col
-    }
-    output
-  }))
+  table_for_widget <- data.frame(index = 0:(nrow(table_for_widget) - 1), table_for_widget)
+
+  # To reduce final size, only output cols we need
+  # Required cols for plotting
+  req_cols <- c("p", "adjusted_p", "uniprot_id",
+                "sig.pval.fct", "sig.FDR.fct",
+                "negLog10rawP", "negLog10adjP",
+                "logFC", "average_intensity",
+                "internal_id_for_brushing", "internal_title_column", "index")
+  # And the user-requested display columns
+  output_cols <- unique(c(req_cols, display.columns))
+
+  table_for_widget <- table_for_widget[,colnames(table_for_widget) %in% output_cols, drop = F]
+
+
   # Counts is a matrix, can just round it
-  counts <- round(counts, digits = 4)
+  counts <- data.frame(round(counts, digits = 4))
 
+  groups_df <- data.frame(group = groups,
+                          sample = colnames(counts))
 
   # Build the data to pass to the htmlwidget
-  xData <- UAMS_buildXYData(table, status, main, display.columns,
-                            anno, counts, status.cols, sample.cols, groups)
+  xData <- list(
+    data = list(
+      table = table_for_widget,
+      cols = display.columns,
+      counts = counts,
+      groups = groups_df,
+      expCols = colnames(groups_df),
+      numUniqueGroups = length(unique(groups_df$group)),
+      statusColours = status.cols,
+      sampleColours = if (is.null(sample.cols)) {-1} else {sample.cols}
+    )
+  )
 
   # make the widget
   widget <- htmlwidgets::createWidget(
@@ -73,38 +91,4 @@ uams_glimmaXY <- function(model_data,
   )
 
   widget
-}
-
-
-UAMS_buildXYData <- function(table, status, main, display.columns, anno, counts,
-                             status.cols, sample.cols, groups) {
-
-  counts <- data.frame(counts)
-  level <- levels(groups)
-  groups <- data.frame(group = groups)
[…]
