### 2023-03-10 4a98399
document internal glimma function
