+++ man/uams_glimmaXY.Rd
@@ -1,0 +1,55 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/glimma_fxns.R
+\name{uams_glimmaXY}
+\alias{uams_glimmaXY}
+\title{Make an HTMLwidget for the interactive report}
+\usage{
+uams_glimmaXY(
+  model_data,
+  counts,
+  groups,
+  anno,
+  display.columns,
+  status.cols,
+  sample.cols,
+  width,
+  height
+)
+}
+\arguments{
+\item{model_data}{The output from the \code{\link{prep_plot_model_data}}
+function, containing statistical results for a single contrast.}
+
+\item{counts}{A matrix or dataframe of the raw or normalized data from the
+data slot of a DAList.}
+
+\item{groups}{A vector of group identities for each sample. Should have
+same length as the number of cols in counts.}
+
+\item{anno}{annotation data, with added p-value columns.}
+
+\item{display.columns}{A vector of columns to display in the output table.}
+
+\item{status.cols}{A vector of colors to use for down regulated, nonDE,
+and upregulated proteins. Must be of length 3.}
+
+\item{sample.cols}{A vector of colors for each sample. Should have the same
+length as groups.}
+
+\item{width}{The width of the interactive report objects, in pixels..}
+
+\item{height}{The height of the interactive report objects, in pixels.}
+}
+\value{
+An HTMLwidget containing our interactive plots and tables
+}
+\description{
+This internal function is called within the .Rmd report template as
+part of the user-facing \code{\link{write_limma_plots}} function. It takes in
+data on statistical results, annotation, and sample intensities (partially
+assembled in \code{\link{write_limma_plots}} and then passed on the to .RMD
+environment), does some further processing, packages
+it into the form needed for our interactive report, and then outputs an
+HTMLwidget for use in the HTML file generated from the .Rmd.
+}
+\keyword{internal}
