+++ inst/report_templates/limma_report_per_contrast.Rmd
@@ -28,7 +28,6 @@ uams_glimmaXY(
   groups = groups,
   sample.cols = unname(colorGroup(groups)),
   status.cols = c("#00bfff", "#858585", "#ff3030"),
-  status = status,
   anno = anno,
   display.columns = cols_to_display,
   width = width,
