+++ inst/htmlwidgets/lib/GlimmaV2/expressionSpec.js
@@ -1,156 +1,155 @@
+// Returns the vega spec for the right side (abundance) plot
+// called "expression" or "exp" across most of the glimma code,
+// which was originally for RNAseq stuff
 function createExpressionSpec(width, height, expColumns, sampleColours, numUniqueGroups)
 {
+  // Add the normalized intensity data to the expColumns,
+  // so it gets included in the tooltip
+  expColumns.push("normalized intensity");
+  // Make the tooltip strings first using the JS function
+  // in lib/GlimmaV2/makeTooltip.js
+  let tooltip = makeVegaTooltip(expColumns);
+  let tooltip_mean = makeVegaTooltip(["group", "mean"]);
 
-    /* need an empty signal if sample.cols argument has been supplied */
-    let samplecols_signal = { "name": "samplecols_active" };
+  return {
+    "$schema": "https://vega.github.io/schema/vega/v5.json",
+    "width": width*0.40,
+    "height": height*0.35,
+    "padding": {"left": 0, "top": 0, "right": 0, "bottom": 10},
+    "autosize": {"type": "fit", "resize": true},
+    "title": { "text": {"signal": "title_signal" }},
+    "signals": [
+      {
+        "name": "title_signal",
+        "value": ""
+      },
+      {
+        "name": "ylim_min",
+        "value": null,
+        "bind": {
+          "input": "number",
+          "class": "ylim_min"
+        }
+      },
+      {
+        "name": "ylim_max",
+        "value": null,
+        "bind": {
+          "input": "number",
+          "class": "ylim_max"
+        }
+      },
+      {
+        "name": "min_count",
+        "value": null
+      },
+      {
+          "name": "max_count",
+          "value": 0
+      },
+      {
+          "name": "min_y",
+          "update": " (ylim_min === null || ylim_min == \"\") ? null : (ylim_min > min_count) ? null : ylim_min"
+      },
+      {
+          "name": "max_y",
+          "update": " (ylim_max < max_count) ? null : ylim_max"
+      },
+      {
+        "name": "mean_linewidth",
+        "value":  ((width*0.4)/(numUniqueGroups + 1)*0.6)*((width*0.4)/(numUniqueGroups + 1)*0.6)
+      }
+    ],
+    "data": [
+      {"name": "table"},
+      {"name": "means",
+       "source" : "table",
+       "transform" : [
+         {"type": "aggregate",
+           "groupby": ["group"],
+           "fields": ["normalized intensity"],
+           "ops": ["mean"],
+           "as": ["mean"]
 
-    /* must match counts term in processExpression */
-    expColumns.push("normalized intensity");
-    let tooltip = makeVegaTooltip(expColumns);
-    let tooltip_mean = makeVegaTooltip(["group", "mean"]);
-
-    return {
-        "$schema": "https://vega.github.io/schema/vega/v5.json",
-        "width": width*0.40,
-        "height": height*0.35,
-        "padding": {"left": 0, "top": 0, "right": 0, "bottom": 10},
-        "autosize": {"type": "fit", "resize": true},
-        "title": { "text": {"signal": "title_signal" }},
-        "signals":
-                [
-                    {
-                        "name": "title_signal",
-                        "value": ""
-                    },
-                    {
-                      "name": "ylim_min",
-                      "value": null,
-                      "bind": {
-                                "input": "number",
-                                "class": "ylim_min"
-                              }
-                    },
-                    {
-                      "name": "ylim_max",
-                      "value": null,
-                      "bind": {
-                                "input": "number",
-                                "class": "ylim_max"
-                              }
-                    },
-                    {
-                      "name": "min_count",
-                      "value": null
-                    },
-                    {
-                        "name": "max_count",
-                        "value": 0
-                    },
-                    {
[…]
