### 2023-03-10 f3bd8cd
restyle XYspecs.js
