+++ inst/htmlwidgets/lib/GlimmaV2/XYSpecs.js
@@ -12,123 +12,123 @@ function createXYSpec(xyData, xyTable, width, height)
     "height": height * 0.35,
     "padding": {"left": 0, "top": 0, "right": 0, "bottom": 10},
     "autosize": {"type": "fit", "resize": true},
-    "signals":
-      [
-        {
-          "name": "click", "value": null,
-          "on": [ {"events": "mousedown", "update": "[datum, now()]" } ]
-        },
-        // Some custom signals
-        // The two main ones are the plot_type and pval_type,
-        // which are bound to selection boxes using vega's selection binding
-        // Once plot type and pval type are chosen, those signals are passed
-        // down to other signals which output the appropriate columns to use for
-        // the axes, axes titles, etc.
-        {
-          "name": "plot_type",
-          "value": "volcano",
-          "bind": {
-            "input": "select",
-            "options": ["volcano", "MD"],
-            "name": "Plot type: ",
-            "style": "width: 200px; margin-bottom: 10px;"
-          }
-        },
-        {
-          "name": "pval_type",
-          "value": "adjusted",
-          "bind": {
-            "input": "select",
-            "options": ["raw", "adjusted"],
-            "name": "P-value type: ",
-            "style": "width: 200px; margin-bottom: 10px;"
-          }
-        },
-        // Set x axis data and title based on plot type
-        // Only two options
-        {
-          "name": "x_axis",
-          "value": "logFC",
-          "on": [
-            {
-             "events": [
-                {"signal": "pval_type"},
-                {"signal": "plot_type"}
-                ],
-              "update": "(plot_type == \"MD\" ? \"average_intensity\" : \"logFC\")"
-            }
-          ]
-        },
-        {
-          "name": "x_axis_title",
-          "value": "log2 fold-change (logFC)",
-          "on": [
-            {
-             "events": [
-                {"signal": "pval_type"},
-                {"signal": "plot_type"}
-                ],
-              "update": "(plot_type == \"MD\" ? \"average intensity\" : \"log2 fold-change (logFC)\")"
-            }
-          ]
-        },
-        // Setting y axis is slightly more complicated, with three options.
-        // vega has slightly limited syntax, can't use a complicated function,
-        // just use nested ternary opertors.
-        {
-          "name": "y_axis",
-          "value": "negLog10adjP",
-           "on": [
-            {
-              "events": [
-                {"signal": "pval_type"},
-                {"signal": "plot_type"}
-                ],
-              "update": "(plot_type == \"MD\" ? \"logFC\" : (pval_type == \"raw\" ? \"negLog10rawP\" : \"negLog10adjP\"))"
-            }
-            ]
-        },
-                {
-          "name": "y_axis_title",
-          "value": "-log10(adjusted P)",
-           "on": [
-            {
-              "events": [
-                {"signal": "pval_type"},
-                {"signal": "plot_type"}
-                ],
-              "update": "(plot_type == \"MD\" ? \"log2 fold-change (logFC)\" : (pval_type == \"raw\" ? \"-log10(raw P)\" : \"-log10(adjusted P)\"))"
-            }
-            ]
-        },
-        // point color scale just based on p-value type
-        {
-          "name": "point_color",
-          "value": "sig.FDR.fct",
-           "on": [
-            {
-              "events": [
-                {"signal": "pval_type"},
-                {"signal": "plot_type"}
-                ],
-              "update": "(pval_type == \"raw\" ? \"sig.pval.fct\" : \"sig.FDR.fct\")"
-            }
-            ]
+    "signals": [
+      {
+        "name": "click", "value": null,
[…]
