### 2023-03-10 72b0e32
minor styling and comments on glimmaXY.js
