### 2023-03-13 #5184 @ByrumLab
HI @MohmedSoudy, Thank you for taking time to review our package.
1. We have updated the paper.md to include a State of the field section.
2. write_qc_report has been updated for readability.
3. write_limma_plots has been updated with a selector to chose the plot type instead of including each as a separate tab. We added a mean line to the point plot to easily see the differences between the groups. The box plot was harder to visualize, especially for projects with only three replicates.

We appreciate your time and helpful suggestions.
