### 2023-03-20 fa16967
fix an error in the examples of filter_samples
