### 2023-03-22 #5184 @MohmedSoudy
- [x] **Repository:** Is the source code for this software available at the [proteoDA](https://github.com/ByrumLab/proteoDA)?
- [x] **Contribution and authorship:** Has the submitting author [ByrumLab](https://github.com/ByrumLab) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
- [x] **Human and animal research:** If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participant's research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.
