### 2023-03-23 #5184 @shahmoradi
I have a few comments for which I'd appreciate the authors' response before I mark the remaining items in the checklist as approved.
1. I am not quite familiar with `testthat` testing framework. Can you clarify how to run the tests? Normally, I'd expect a README file to exist in the appropriate folder that describes the steps to run the tests.
2. I cannot find any coverage reports in the repository. Unless I missed the report, a test suite without a code coverage report would be of minimal value. An example R package that could help generate coverage reports is `covr`.
3. I agree with @MohmedSoudy that the "State of the field" can be improved. I see the authors' response, but I cannot find the improvements in the most recent PDF compilation of the paper.
4. I cannot find "Community guidelines" in the repository. If one is already in the repository, please guide me to its location. See the checklist above for more information on "Community guidelines".
