### 2023-03-24 cdf4b0b
add R CMD check github action and badge
