+++ .github/workflows/R-CMD-check.yaml
@@ -40,6 +40,16 @@ jobs:
           extra-packages: any::rcmdcheck
           needs: check
 
+      - name: Manually install preprocessCore
+        run: |
+          install.packages(
+            "preprocessCore",
+            repos = BiocManager::repositories(),
+            type = "source",
+            configure.args = "--disable-threading"
+          )
+        shell: Rscript {0}
+
       - uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
