### 2023-03-24 c20e91f
more GHA fixes
stop testing against R devel on ubuntu for now (bioconductor is not up to r 4.4 yet), try specifying ggtree and ggplot version to avoid dependency issues
