+++ tests/testthat/test-normalization_report.R
@@ -105,3 +105,25 @@ test_that("write_norm_report creates proper output file", {
   expect_true(file.exists("tempdirfortesting/temp.pdf"))
 
 })
+
+
+test_that("write_norm_report creates proper output file when using ggrastr", {
+  # default filenames
+  input <- readRDS(test_path("fixtures", "norm_report_input.rds"))
+  on.exit(unlink("normalization_report.pdf"), add = T)
+
+  skip_if_not_installed("ggrastr")
+  suppressMessages(write_norm_report(input, grouping_column = "group", use_ggrastr = T, overwrite = T))
+  expect_true(file.exists("normalization_report.pdf"))
+})
+
+
+test_that("write_norm_report creates proper output file when using suppressing zoom legend", {
+  # default filenames
+  input <- readRDS(test_path("fixtures", "norm_report_input.rds"))
+  on.exit(unlink("normalization_report.pdf"), add = T)
+
+  skip_if_not_installed("ggrastr")
+  suppressMessages(write_norm_report(input, grouping_column = "group", suppress_zoom_legend = T, overwrite = T))
+  expect_true(file.exists("normalization_report.pdf"))
+})
