### 2023-03-29 12a3d01
update badges to work on ByrumLab, run GHA tests on PR
