+++ .github/workflows/R-CMD-check.yaml
@@ -1,6 +1,6 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
-on: [push]
+on: [pull_request]
 
 name: R-CMD-check
 
