### 2023-03-29 ba5b824
custom install of preprocessCore for test coverage
and modify when GHA are done: all pushes and PRs on main branches
