+++ .github/workflows/test-coverage.yaml
@@ -1,6 +1,12 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
-on: [pull_request]
+on:
+  push:
+    branches: [main, master]
+  pull_request:
+    branches: [main, master]
+
+
 name: test-coverage
 
 jobs:
@@ -20,6 +26,16 @@ jobs:
         with:
           extra-packages: any::covr
           needs: coverage
+
+      - name: Manually install preprocessCore
+        run: |
+          install.packages(
+            "preprocessCore",
+            repos = BiocManager::repositories(),
+            type = "source",
+            configure.args = "--disable-threading"
+          )
+        shell: Rscript {0}
 
       - name: Test coverage
         run: |
