### 2023-03-30 #5184 @ByrumLab
Thank you @shahmoradi, we appreciate the review! We've made some changes to the package in response to your review:

1. The `testthat` framework is very commonly used across R packages, but it does have some differences to the way tests are distributed with other software packages and languages. In general, R packages do not have their tests bundled with them, so users cannot re-run `testthat` locally. See, for example, popular R packages like [dplyr](https://github.com/tidyverse/dplyr/) and [ggplot2](https://github.com/tidyverse/ggplot2/), which do not bundle their tests for distribution (that is, the tests are not in the `inst/` directory). Instead, `testthat` runs package tests locally during development and during automated package checking with `R CMD check`. We have set up some GitHub actions to automatically run `R CMD check` against the latest R releases for the Mac, Windows, and Unix/Ubuntu operating systems, with the results displayed in a badge on the GitHub README. These tests are also run when packages are submitted to CRAN, which we plan to do upon acceptance.

2. We have used the R package you suggested to generate a code coverage report: it can be found by clicking on the codecov badge in our README. Currently, we have ~100% test coverage.

4. We have added community guidelines to our README, and have included a contributor code of conduct (CODE_OF_CONDUCT.md).

We hope these changes address your concerns.
