+++ CITATION.cff
@@ -28,3 +28,5 @@ authors:
 title: ByrumLab/proteoDA: Initial Release
 version: 1.0.0
 date-released: 2023-05-23
+doi: 10.5281/zenodo.7962306
+url: "https://doi.org/10.5281/zenodo.7962306"
