### 2023-05-24 #5184 @jmschrei
@ByrumLab would you mind fixing these two invalid DOIs?
