### 2023-05-24 a8e66b8
Update references.bib
Fixing invalid doi references as per this [review](https://github.com/openjournals/joss-reviews/issues/5184#issuecomment-1561546600).
