+++ paper/references.bib
@@ -92,7 +92,7 @@ volume = {41},
 number = {2},
 pages = {1800042},
 keywords = {mass spectrometry, MRM, PRM, SRM, targeted proteomics},
-doi = {https://doi.org/10.1002/bies.201800042},
+doi = {10.1002/bies.201800042},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/bies.201800042},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/bies.201800042},
 abstract = {While mass spectrometry (MS)-based quantification of small molecules has been successfully used for decades, targeted MS has only recently been used by the proteomics community to investigate clinical questions such as biomarker verification and validation. Targeted MS holds the promise of a paradigm shift in the quantitative determination of proteins. Nevertheless, targeted quantitative proteomics requires improvisation in making sample processing, instruments, and data analysis more accessible. In the backdrop of the genomic era reaching its zenith, certain questions arise: is the proteomic era about to come? If we are at the beginning of a new future for protein quantification, are we prepared to incorporate targeted proteomics at the benchside for basic research and at the bedside for the good of patients? Here, an overview of the knowledge required to perform targeted proteomics as well as its applications is provided. A special emphasis is placed on upcoming areas such as peptidomics, proteoform research, and mass spectrometry imaging, where the utilization of targeted proteomics is expected to bring forth new avenues. The limitations associated with the acceptance of this technique for mainstream usage are also highlighted. Also see the video abstract here https://youtu.be/mieB47B8gZw.},
@@ -117,7 +117,7 @@ volume = {22},
 number = {8},
 pages = {2100103},
 keywords = {bottom-up proteomics, data visualization, open science, science communication},
-doi = {https://doi.org/10.1002/pmic.202100103},
+doi = {10.1002/pmic.202100103},
 url = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/abs/10.1002/pmic.202100103},
 eprint = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/pdf/10.1002/pmic.202100103},
 abstract = {Abstract Mass-spectrometry based bottom-up proteomics is the main method to analyze proteomes comprehensively and the rapid evolution of instrumentation and data analysis has made the technology widely available. Data visualization is an integral part of the analysis process and it is crucial for the communication of results. This is a major challenge due to the immense complexity of MS data. In this review, we provide an overview of commonly used visualizations, starting with raw data of traditional and novel MS technologies, then basic peptide and protein level analyses, and finally visualization of highly complex datasets and networks. We specifically provide guidance on how to critically interpret and discuss the multitude of different proteomics data visualizations. Furthermore, we highlight Python-based libraries and other open science tools that can be applied for independent and transparent generation of customized visualizations. To further encourage programmatic data visualization, we provide the Python code used to generate all data figures in this review on GitHub (https://github.com/MannLabs/ProteomicsVisualization).},
