### 2023-05-24 2faa931
author name correction
Based on this [proof](https://github.com/openjournals/joss-reviews/issues/5184#issuecomment-1561545832), it became apparent that the following author name needes to be changed from : "II Michael S. Robeson"  to: "Michael S. Robeson II"
