### 2023-05-24 #5184 @mikerobeson
Fixed affiliation 2 from "Biomedical Imaging" to "Biomedical Informatics". Committed changes.
