### 2023-05-25 ee8dd5a
fix typo
Correcting the typo as referenced [here](https://github.com/openjournals/joss-reviews/issues/5184#issuecomment-1562529797).

Changed "reserchers" to "researchers".
