### 2023-05-30 #5184 @Kevin-Mattheus-Moerman
> @openjournals/dev this DOI is taking very long to resolve point_up

Resolves now
